Hangman Game
General Hangman game consists of a limited number of predefined guessing occurrences. But in this hangman game, the number of chances for each execution changes in random, and the words have been shuffled and selected in random for each attempt. The words have been categorized as continent names. During each guess, the number of guesses and number of letters left in the word has been displayed.  After completion of each game, a prompt has been designed to ask the user whether the user wants to play again or exit the game.